+++ internal/client/job_definition.go
@@ -6,8 +6,8 @@ import (
 	"terraform-provider-trocco/internal/client/entity"
 	jobDefinitionEntities "terraform-provider-trocco/internal/client/entity/job_definition"
 	filterEntities "terraform-provider-trocco/internal/client/entity/job_definition/filter"
-	inputOptionEntitites "terraform-provider-trocco/internal/client/entity/job_definition/input_options"
-	outputOptionEntitites "terraform-provider-trocco/internal/client/entity/job_definition/output_options"
+	inputOptionEntitites "terraform-provider-trocco/internal/client/entity/job_definition/input_option"
+	outputOptionEntitites "terraform-provider-trocco/internal/client/entity/job_definition/output_option"
 	"terraform-provider-trocco/internal/client/parameter"
 	"terraform-provider-trocco/internal/client/parameter/job_definitions"
 	filterParameters "terraform-provider-trocco/internal/client/parameter/job_definitions/filter"
